pyproject.toml
[tool.poetry]
name = "product catalog"
version = "0.1.0"
description = ""
authors = ["vipdev"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.12"
django = "^5.1"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
python-dotenv = "^1.0.1"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
